Intermediate commit for logic implementation.
