--- src/main/java/components/Bar.java
@@ -1,8 +1,0 @@
-package components;
-
-public class Bar extends Slot{
-    public Bar(String name, int odd) {
-        super(name, odd);
-    }
-
-}
