--- src/main/java/components/Cherry.java
@@ -1,7 +1,0 @@
-package components;
-
-public class Cherry extends Slot {
-    public Cherry(String name, int odd) {
-        super(name, odd);
-    }
-}
