--- src/main/java/components/Diamond.java
@@ -1,9 +1,0 @@
-package components;
-
-public class Diamond extends Slot {
-    public Diamond(String name, int odd) {
-        super(name, odd);
-    }
-
-
-}
