--- src/main/java/components/Heart.java
@@ -1,7 +1,0 @@
-package components;
-
-public class Heart extends Slot {
-    public Heart(String name, int odd) {
-        super(name, odd);
-    }
-}
