--- src/main/java/components/Seven.java
@@ -1,9 +1,0 @@
-package components;
-
-public class Seven extends Slot {
-    public Seven(String name, int odd) {
-        super(name, odd);
-    }
-
-
-}
