--- src/main/java/components/Shamrock.java
@@ -1,7 +1,0 @@
-package components;
-
-public class Shamrock extends Slot{
-    public Shamrock(String name, int odd) {
-        super(name, odd);
-    }
-}
