--- src/main/java/components/Watermelon.java
@@ -1,7 +1,0 @@
-package components;
-
-public class Watermelon extends Slot{
-    public Watermelon(String name, int odd) {
-        super(name, odd);
-    }
-}
