+++ src/main/java/core/components/slot/Bar.java
@@ -1,0 +1,10 @@
+package core.components.slot;
+
+import core.components.AbstractSlot;
+
+public class Bar extends AbstractSlot {
+    public Bar(String name, int odd) {
+        super(name, odd);
+    }
+
+}
