+++ src/main/java/core/components/slot/Cherry.java
@@ -1,0 +1,9 @@
+package core.components.slot;
+
+import core.components.AbstractSlot;
+
+public class Cherry extends AbstractSlot {
+    public Cherry(String name, int odd) {
+        super(name, odd);
+    }
+}
