+++ src/main/java/core/components/slot/Diamond.java
@@ -1,0 +1,11 @@
+package core.components.slot;
+
+import core.components.AbstractSlot;
+
+public class Diamond extends AbstractSlot {
+    public Diamond(String name, int odd) {
+        super(name, odd);
+    }
+
+
+}
