+++ src/main/java/core/components/slot/Heart.java
@@ -1,0 +1,9 @@
+package core.components.slot;
+
+import core.components.AbstractSlot;
+
+public class Heart extends AbstractSlot {
+    public Heart(String name, int odd) {
+        super(name, odd);
+    }
+}
