+++ src/main/java/core/components/slot/Seven.java
@@ -1,0 +1,11 @@
+package core.components.slot;
+
+import core.components.AbstractSlot;
+
+public class Seven extends AbstractSlot {
+    public Seven(String name, int odd) {
+        super(name, odd);
+    }
+
+
+}
