+++ src/main/java/core/components/slot/Shamrock.java
@@ -1,0 +1,9 @@
+package core.components.slot;
+
+import core.components.AbstractSlot;
+
+public class Shamrock extends AbstractSlot {
+    public Shamrock(String name, int odd) {
+        super(name, odd);
+    }
+}
