+++ src/main/java/core/components/slot/Watermelon.java
@@ -1,0 +1,9 @@
+package core.components.slot;
+
+import core.components.AbstractSlot;
+
+public class Watermelon extends AbstractSlot {
+    public Watermelon(String name, int odd) {
+        super(name, odd);
+    }
+}
